+++ src/build-rl/game_rl.c
@@ -338,6 +338,14 @@ Gui_Button floor_buttons[GUI_FLOOR_GROUP_BUTTONS_MAX+1] = {
         .label_len = ARRAY_SIZE("Artifacts")-1,
         .box_color = GUI_FLOOR_GROUP_BOX_COLOR,
         .text_color = GUI_FLOOR_GROUP_TEXT_COLOR,
+    },
+    [BUTTON_FLOOR_SAVE] = {
+        .on = false,
+        .state = BUTTON_NORMAL,
+        .label = "Save",
+        .label_len = ARRAY_SIZE("Save")-1,
+        .box_color = GUI_FLOOR_GROUP_BOX_COLOR,
+        .text_color = GUI_FLOOR_GROUP_TEXT_COLOR,
     }
 };
 Gui_Button_Group floor_buttons_group = {
@@ -1599,19 +1607,14 @@ void update_GameScreen(Gui_State* gui_state, Floor** current_floor, Path** game_
         Gui_Button_Group group = gui_state->floor_buttons;
         for (Gui_Floor_Group_Button_Index i = 0; i < group.len; i++) {
             group.buttons[i].on = false;
-            switch (i) {
-            case BUTTON_ARTIFACTS: {
-                int bt_w = gui_state->gameScreenWidth*0.2f;
-                int bt_h = gui_state->gameScreenHeight*0.1f;
-                group.buttons[i].r = (Rectangle) {
-                    .x = gui_state->gameScreenWidth - (bt_w*2),
-                    .y = (gui_state->gameScreenHeight - bt_h)*0.5f,
-                    .width = bt_w,
-                    .height = bt_h
-                };
-            }
-            break;
-            }
+            int bt_w = gui_state->gameScreenWidth*0.2f;
+            int bt_h = gui_state->gameScreenHeight*0.1f;
+            group.buttons[i].r = (Rectangle) {
+                .x = gui_state->gameScreenWidth - (bt_w*2),
+                .y = (gui_state->gameScreenHeight - bt_h)*0.5f + (i * bt_h),
+                .width = bt_w,
+                .height = bt_h
+            };
         }
         for (Gui_Floor_Group_Button_Index i = 0; i < group.len; i++) {
             if (CheckCollisionPointRec(gui_state->virtualMouse, group.buttons[i].r)) {
@@ -1633,11 +1636,35 @@ void update_GameScreen(Gui_State* gui_state, Floor** current_floor, Path** game_
                     gui_state->currentScreen = ARTIFACTS_VIEW;
                 }
                 break;
+                case BUTTON_FLOOR_SAVE: {
+                    // Save
+                    {
+                        Enemy *dummy_enemy = NULL;
+                        Boss *dummy_boss = NULL;
+                        FILE *dummy_savefile = NULL;
+                        Rectangle rect = {0};
+                        RingaBuf *dummy_rb = NULL;
+                        foeTurnOption_OP dummy_foe_op = FOE_OP_INVALID;
+                        skillType dummy_skill_pick = -1;
+                        //Declare turnOP_args
+                        turnOP_args *args =
+                            init_turnOP_args(*gamestate, *player, *game_path, *current_room, load_info, dummy_enemy,
+                                             dummy_boss, dummy_savefile, &rect, *floor_kls,
+                                             dummy_foe_op, dummy_skill_pick, dummy_rb);
+
+                        log_tag("debug_log.txt", "[DEBUG]", "%s():    Skipping preparing save file path", __func__);
+                        turnOP(turnOP_from_turnOption(SAVE), args, default_kls, *floor_kls);
+                        log_tag("debug_log.txt", "[DEBUG]", "%s():    Done save.", __func__);
+                        log_tag("debug_log.txt", "[DEBUG]", "%s():    G_RNG_ADVANCEMENTS == {%" PRId64 "}", __func__, G_RNG_ADVANCEMENTS);
+                        log_tag("debug_log.txt", "[DEBUG]", "%s():    Seed: {%s}", __func__, (*game_path)->seed);
+                    }
+                }
+                break;
                 }
             }
         }
 
-        if (IsKeyPressed(KEY_UP)) {
+        if (IsKeyPressed(KEY_UP) || IsKeyPressed(KEY_W)) {
             step_floor(*current_floor, current_x,
                        current_y, KEY_UP);
             (*player)->floor_x = *current_x;
@@ -1648,7 +1675,7 @@ void update_GameScreen(Gui_State* gui_state, Floor** current_floor, Path** game_
                 break;
             }
         }
-        if (IsKeyPressed(KEY_DOWN)) {
+        if (IsKeyPressed(KEY_DOWN) || IsKeyPressed(KEY_S)) {
             step_floor(*current_floor, current_x,
                        current_y, KEY_DOWN);
             (*player)->floor_x = *current_x;
@@ -1659,7 +1686,7 @@ void update_GameScreen(Gui_State* gui_state, Floor** current_floor, Path** game_
                 break;
             }
         }
-        if (IsKeyPressed(KEY_LEFT)) {
+        if (IsKeyPressed(KEY_LEFT) || IsKeyPressed(KEY_A)) {
             step_floor(*current_floor, current_x,
                        current_y, KEY_LEFT);
             (*player)->floor_x = *current_x;
@@ -1670,7 +1697,7 @@ void update_GameScreen(Gui_State* gui_state, Floor** current_floor, Path** game_
                 break;
             }
         }
-        if (IsKeyPressed(KEY_RIGHT)) {
+        if (IsKeyPressed(KEY_RIGHT) || IsKeyPressed(KEY_D)) {
             step_floor(*current_floor, current_x,
                        current_y, KEY_RIGHT);
             (*player)->floor_x = *current_x;
@@ -2269,6 +2296,7 @@ void update_GameScreen(Gui_State* gui_state, Floor** current_floor, Path** game_
                         group->buttons[j].state = BUTTON_NORMAL;
                     }
                     if (group->buttons[j].on) {
+                        group->buttons[j].on = false;
                         switch (i) {
                         case SHOP_LAYOUT_EQUIPS_GROUP: {
                             if ((*player)->equipsBagOccupiedSlots >= EQUIPSBAGSIZE) {
@@ -2564,10 +2592,10 @@ void update_GameScreen(Gui_State* gui_state, Floor** current_floor, Path** game_
         if (gui_state->selectedIndex >= (*player)->equipsBagOccupiedSlots) {
             gui_state->selectedIndex = 0;
         }
-        if (IsKeyPressed(KEY_DOWN)) {
+        if (IsKeyPressed(KEY_DOWN || IsKeyPressed(KEY_S))) {
             gui_state->selectedIndex += 1;
         }
-        if (IsKeyPressed(KEY_UP)) {
+        if (IsKeyPressed(KEY_UP) || IsKeyPressed(KEY_W)) {
             if (gui_state->selectedIndex > 0) gui_state->selectedIndex -= 1;
         }
         if (IsKeyPressed(KEY_ENTER)) {
@@ -2618,10 +2646,10 @@ void update_GameScreen(Gui_State* gui_state, Floor** current_floor, Path** game_
         if (gui_state->selectedIndex >= CONSUMABLESMAX+1) {
             gui_state->selectedIndex = 0;
         }
-        if (IsKeyPressed(KEY_DOWN)) {
+        if (IsKeyPressed(KEY_DOWN) || IsKeyPressed(KEY_S)) {
             gui_state->selectedIndex = (gui_state->selectedIndex + 1) % (CONSUMABLESMAX+1);
         }
-        if (IsKeyPressed(KEY_UP)) {
+        if (IsKeyPressed(KEY_UP) || IsKeyPressed(KEY_W)) {
             if (gui_state->selectedIndex > 0) gui_state->selectedIndex -= 1;
         }
         if (IsKeyPressed(KEY_ENTER)) {
@@ -2652,10 +2680,10 @@ void update_GameScreen(Gui_State* gui_state, Floor** current_floor, Path** game_
         if (gui_state->selectedIndex >= ARTIFACTSMAX+1) {
             gui_state->selectedIndex = 0;
         }
-        if (IsKeyPressed(KEY_DOWN)) {
+        if (IsKeyPressed(KEY_DOWN) || IsKeyPressed(KEY_S)) {
             gui_state->selectedIndex = (gui_state->selectedIndex + 1) % (ARTIFACTSMAX+1);
         }
-        if (IsKeyPressed(KEY_UP)) {
+        if (IsKeyPressed(KEY_UP) || IsKeyPressed(KEY_W)) {
             if (gui_state->selectedIndex > 0) gui_state->selectedIndex -= 1;
         }
     }